[CHORE] TodayNavigationSpy를 Dummy로 네이밍 변경 (#197)
